+++ hutool-crypto/src/test/java/cn/hutool/crypto/digest/OTPTest.java
@@ -1,7 +1,6 @@
-package cn.hutool.crypto.test.digest;
+package cn.hutool.crypto.digest;
 
 import cn.hutool.core.codec.Base32;
-import cn.hutool.crypto.digest.HmacAlgorithm;
 import cn.hutool.crypto.digest.otp.HOTP;
 import cn.hutool.crypto.digest.otp.TOTP;
 import org.junit.Assert;